gemini-ai/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai

groq
langchain-groq
langchain_community
pypdf
google-cloud-aiplatform>=1.38
numpy==1.24.3
pillow

beautifulsoup4 
pandas 
nltk 
textblob 
plotly 
requests

selenium
webdriver-manager